requirements.txt
requests>=2.31
beautifulsoup4>=4.12
python-dotenv>=1.0
#Note!!! Ensure you have the necessary permissions to scrape on the target websites. Always respect robots.txt and terms of service.

